[coin] Add batch coin transfer payload
